Initial commit
This is configured for my use, just how I like it. Feel free to mess around with it and change it to suit you. Open tree in one window, dmesg in another, cmatrix in a third and look like the worlds most badass hacker. Don't forget to have your hood up and ray-bans on for the full effect.